+++ src/Value/NoRating.php
@@ -42,7 +42,7 @@ final class NoRating extends ValueAbstract implements RatingInterface
      */
     public function sameValueAs(ValueInterface $object): bool
     {
-        return $this->sameValueTypeAs($object);
+        return $object instanceof self;
     }
 
     /**